snaksnak

+++ application/controllers/Employee_reports.php
@@ -180,7 +180,7 @@ class Employee_reports extends HR_Controller
 		}
 		else if(isset($input['remove'])){
 			$report = $this->reports->get(['id' => $input['id']]);
-			if(!$report['status'] || ($report['status'] $this->session->userdata('account_type')=='ad')){
+			if(!$report['status'] || ($report['status'] && $this->session->userdata('account_type')=='ad')){
 				$this->reports->update(['image' => NULL, 'condition' => ['id' => $input['id']]]);
 				unlink('./assets/img/reports/'.$report['image']);
